fix: 위도 경도 타입형 double 변경

+++ src/main/java/Butlers/Ticat/weather/entity/Region.java
@@ -22,14 +22,14 @@ public class Region {
     @Column(name = "region_child")
     private String childRegion;
 
-    private int nx;
+    private double nx;
 
-    private int ny;
+    private double ny;
 
     @Embedded
     private Weather weather;
 
-    public Region(Long id, String parentRegion, String childRegion, int nx, int ny) {
+    public Region(Long id, String parentRegion, String childRegion, double nx, double ny) {
         this.id = id;
         this.parentRegion = parentRegion;
         this.childRegion = childRegion;
